apres un bug de 8h j'ai réussi a faire tourner le shell wha la galère et c a tourne sur vrai hardware donc nice

+++ kernel/serial/com1.h
@@ -5,6 +5,6 @@ void serial_init();
 void serial_putchar(char c);
 void serial_print(char *str);
 void serial_println(char *str);
-void serial_print_hex(uint32_t n); 
+void serial_print_hex(uint64_t n); 
 
 #endif
